refactor(sourcegen): extension files, enum model size
